+++ lib/src/rules/all_rules.dart
@@ -132,6 +132,7 @@ export 'platforms/ios_capabilities_permissions_rules.dart';
 export 'platforms/ios_platform_lifecycle_rules.dart';
 export 'platforms/ios_ui_security_rules.dart';
 export 'platforms/linux_rules.dart';
+export 'platforms/method_channel_rules.dart';
 export 'platforms/macos_rules.dart';
 export 'platforms/plugin_platform_support_rules.dart';
 export 'platforms/web_rules.dart';